% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo.bull.R
\name{hanwoo.bull}
\alias{hanwoo.bull}
\title{hanwoo.bull}
\usage{
hanwoo.bull(KPN)
}
\arguments{
\item{KPN}{KPN number of a bull you get the inform.}
}
\description{
This function scraping the bull information of Hanwoo from data.go.kr.
}
\examples{
hanwoo.bull(1080)

}
\keyword{Hanwoo}
\keyword{bull}
