% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanwoo.info.R
\name{hanwoo.info}
\alias{hanwoo.info}
\title{hanwoo.info}
\usage{
hanwoo.info(cattle)
}
\arguments{
\item{cattle}{Number of cattle you get the inform.}
}
\description{
This function scraping the information of Hanwoo from data.go.kr. Please get your API key and request for applicate at data.go.kr.
}
\examples{
hanwoo.info("002083191603")
}
\keyword{Hanwoo}
